Add support for copying file data by character and by line
The code changes introduce two new functions, `copiaFileC` and `copiaFileR`, which allow for copying data from one file to another. The `copiaFileC` function copies the data character by character, while the `copiaFileR` function copies the data line by line. These functions are implemented in the file `Es96-F.cpp`.

This commit also includes the addition of three new text files: `in.txt`, `out.txt`, and `outR.txt`. These files will be used to test the functionality of the `copiaFileC` and `copiaFileR` functions.

The purpose of this commit is to provide the functionality to copy file data using different methods and to prepare for testing the results.